processEachFile: ENH: Also handle non-existing FILE
Both before and after processing. If the file gets created or deleted, that's a change. If it stays non-existing, it's no change.

As cmp would "exit 2" on a non-existing file, and the digests complain about a missing file, too, we need a special case that uses the null device for the copyFilespec and an empty digest value.