pipethrough[1]: ENH: Also handle non-existing FILE
Here only before processing. Due to the redirect of the command execution into the tempfile, a non-existing FILE will always get created (as long as the command itself succeeds). If the file gets created or truncated, that's a change. If it did not exist before and is created as an empty file, that's no change.

As cmp would "exit 2" on a non-existing file, and the digests complain about a missing file, too, we need a special case that uses the null device.