ENH: Allow {ii} alternative to {I} that includes the (non-recursive) file glob
With --exec, all COMMAND arguments are quoted, so it's not possible to pass a "{I}/*" glob argument - --command would need to be used.
As it's quite common that commands take individual FILE arguments and not a directory, and the escaping with --command is cumbersome, let's offer a non-recursive (but including hidden files) glob variant in the form of {ii}.
Note: For the output dir, this does not apply; we need to pass an actual directory there as output files are not necessarily named the same (then we wouldn't need this command and could just iterate on the original dir) and there's not necessarily a 1:1 mapping between input and output files.